max_shrinkseconds: float->int

+++ hypothesis-python/src/hypothesis/_settings.py
@@ -156,7 +156,7 @@ class settings(metaclass=settingsMeta):
         suppress_health_check: Collection["HealthCheck"] = not_set,  # type: ignore
         deadline: Union[int, float, datetime.timedelta, None] = not_set,  # type: ignore
         print_blob: bool = not_set,  # type: ignore
-        max_shrink_seconds: float = not_set,  # type:ignore
+        max_shrink_seconds: int = not_set,  # type:ignore
     ) -> None:
         if parent is not None:
             check_type(settings, parent, "parent")
